requirements.txt
black==24.8.0
celery==5.5.3
django-celery-beat==2.8.1
django-phonenumber-field==8.0.0
django-smsru==1.0.4
django==5.1.3
phonenumbers==8.13.50
pillow==11.0.0
psycopg2-binary==2.9.10
python-telegram-bot==22.0
requests==2.32.5
uvicorn[standard]==0.32.0

pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1